refactor: move includeTypes to BaseTypeGen
